+++ src/OpenTelemetry.Instrumentation.SqlClient/Implementation/SqlEventSourceListener.netfx.cs
@@ -140,7 +140,7 @@ internal sealed class SqlEventSourceListener : EventListener
 
         var dataSource = (string)eventData.Payload[1];
         var databaseName = (string)eventData.Payload[2];
-        var startTags = SqlActivitySourceHelper.GetTagListFromConnectionInfo(dataSource, databaseName, out var activityName);
+        var startTags = SqlTelemetryHelper.GetTagListFromConnectionInfo(dataSource, databaseName, out var activityName);
         var commandText = (string)eventData.Payload[3];
         if (!string.IsNullOrEmpty(commandText))
         {
@@ -153,7 +153,7 @@ internal sealed class SqlEventSourceListener : EventListener
             }
         }
 
-        var activity = SqlActivitySourceHelper.ActivitySource.StartActivity(
+        var activity = SqlTelemetryHelper.ActivitySource.StartActivity(
             activityName,
             ActivityKind.Client,
             default(ActivityContext),
@@ -194,7 +194,7 @@ internal sealed class SqlEventSourceListener : EventListener
         // Ensure any activity that may exist due to ActivitySource.AddActivityListener()
         // is stopped regardless of whether we're doing metrics and/or tracing.
         // See https://github.com/open-telemetry/opentelemetry-dotnet-contrib/issues/3033.
-        var sqlActivity = currentActivity?.Source == SqlActivitySourceHelper.ActivitySource ? currentActivity : null;
+        var sqlActivity = currentActivity?.Source == SqlTelemetryHelper.ActivitySource ? currentActivity : null;
 
         // If we're only collecting metrics, then we don't want to modify the activity
         var traceActivity =
@@ -242,7 +242,7 @@ internal sealed class SqlEventSourceListener : EventListener
 
         if (activity != null && activity.IsAllDataRequested)
         {
-            foreach (var name in SqlActivitySourceHelper.SharedTagNames)
+            foreach (var name in SqlTelemetryHelper.SharedTagNames)
             {
                 var value = activity.GetTagItem(name);
                 if (value != null)
@@ -253,7 +253,7 @@ internal sealed class SqlEventSourceListener : EventListener
         }
         else
         {
-            tags.Add(SemanticConventions.AttributeDbSystemName, SqlActivitySourceHelper.MicrosoftSqlServerDbSystemName);
+            tags.Add(SemanticConventions.AttributeDbSystemName, SqlTelemetryHelper.MicrosoftSqlServerDbSystemName);
 
             var (hasError, errorNumber, exceptionType) = ExtractErrorFromEvent(eventData);
 
@@ -268,8 +268,8 @@ internal sealed class SqlEventSourceListener : EventListener
         }
 
         var duration = activity?.Duration.TotalSeconds
-            ?? SqlActivitySourceHelper.CalculateDurationFromTimestamp(this.beginTimestamp.Value);
-        SqlActivitySourceHelper.DbClientOperationDuration.Record(duration, tags);
+            ?? SqlTelemetryHelper.CalculateDurationFromTimestamp(this.beginTimestamp.Value);
+        SqlTelemetryHelper.DbClientOperationDuration.Record(duration, tags);
     }
 }
 #endif